+++ open_embedded_stable_ati/oe_at91sam/recipes/ati/files/modules/rf_conn/fasit/rf.h
@@ -37,7 +37,7 @@ void EnQueue(queue_t *M, void *ptr,int count);
 
 int QueuePtype(queue_t *M);
 
-#define Queue_Depth(M) ((M)->tail - (M)->head)
+#define Queue_Depth(M) ((int)((M)->tail - (M)->head))
 
 void print_verbosity_bits(void);
